feat: agregar interfaz de notas, saldar deuda técnica del back

+++ apuntarservidor/src/main/java/com/apuntar/apuntarservidor/aplicacion/puertos/NivelAcademicoRepositoryPort.java
@@ -7,7 +7,7 @@ import com.apuntar.apuntarservidor.dominio.modelos.NivelAcademico;
 
 public interface NivelAcademicoRepositoryPort {
     List<NivelAcademico> findAll();
-    Optional<NivelAcademico> findById(Long id);
+    Optional<NivelAcademico> findById(Integer id);
 
     Optional<NivelAcademico> findByDescripcion(String descripcion);
