+++ apuntarservidor/src/main/java/com/apuntar/apuntarservidor/aplicacion/servicios/PrioridadService.java
@@ -19,7 +19,7 @@ public class PrioridadService {
         return prioridadRepository.findAll();
     }
 
-    public Optional<Prioridad> obtenerPorId(Long id){
+    public Optional<Prioridad> obtenerPorId(Integer id){
         return prioridadRepository.findById(id);
     }
